Separating even more classes into new files. CMakeLists arg substitution.
