+++ moveit_experimental/jog_arm/launch/spacenav_cpp.launch
@@ -7,7 +7,7 @@
 
   <node name="spacenav_node" pkg="spacenav_node" type="spacenav_node" />
 
-  <node name="spacenav_to_twist" pkg="moveit_experimental" type="spacenav_to_twist" output="screen" />
+  <node name="spacenav_to_twist" pkg="moveit_experimental" type="jog_arm_spacenav_to_twist" output="screen" />
 
   <node name="jog_arm_server" pkg="moveit_experimental" type="jog_arm_server" output="screen" >
     <param name="parameter_ns" type="string" value="jog_arm_server" />
