+++ moveit_experimental/jog_arm/src/jog_arm/teleop_examples/spacenav_to_twist.cpp
@@ -5,7 +5,6 @@
 
 namespace jog_arm
 {
-
 static const int NUM_SPINNERS = 1;
 static const int QUEUE_LENGTH = 1;
